refactor(tests): runs CLI in-process rather than in a sub-process
This speeds up the tests a fair amount, but the downside is that global side-effects persist between tests. To work around this, I wrote a class called `ProcessSnapshot` that saves and restores required files, require extensions, and process events. This is somewhat of a hack, but it seems to work.

It only required one change to the CLI tests. Instead of uncaught exceptions being written to stderr they now bubble up to the caller of `runCodemodCLI`. This might be slightly better but is basically a neutral change.