feat: add `matchPath` to enable working with `NodePath` objects
Doing manipulation of ASTs with `NodePath` objects is significantly easier than manually modifying the raw node objects in some cases, especially removing a node from an array.

To facilitate this, we need to keep track of the AST key path for our captured values. This means that each matcher, if it delegates to child matchers, needs to pass the key path along and possibly add to it.