fix(types): improve the types of capturing matchers
Prior to this capturing matchers were parameterized on only one type. This type was used for both the matched value type and the captured value type. This worked fine for the basic `capture` matcher, but does not apply to the `containerOf` matcher. This splits that type into two: a matched value type and a captured value type, which allows proper type handling when `containerOf` is a child of another matcher.