feat: allow assigning options to plugin positionally
With this users can do e.g. `codemod -p my-plugin -o @opts1.json -p my-plugin2 -o @opts2.json …` and have `opts1.json` be used for `my-plugin`'s options and `opts2.json` be used for `my-plugin2`'s options. Previously users had to specify the name of the plugin (or its file basename) when assigning options. Now the CLI can infer it based on argument position.

This preserves the feature of using the declared plugin name, but fails gracefully if doing so throws an exception. The downside is that this feature can, if the plugin instantiation throws, silently fail for reasons that are not clear. I think this trade-off is probably worth it.

Closes #686