fix(matchers): ensure package types work
The built package had types with an import specifier containing `../../src/matchers`. However, since `src` was the build root directory it does not appear in the final output structure. This rendered the import specifier invalid, breaking the types for the `match` and `matchPath` utilities. The fix was to use the correct shortest relative path.